# SDC/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SDC - Sistema de Defesa Civil API",
        "description": "API RESTful escalável para 100k+ usuários simultâneos.\r\n *                  Suporta webhooks, rate limiting inteligente, processamento assíncrono via Redis,\r\n *                  e múltiplos níveis de priorização de requisições.",
        "contact": {
            "name": "SDC API Support",
            "email": "api@sdc.gov.br"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://sdc.gov.br/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.sdc.gov.br",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/v1/webhooks/receive": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Receber webhook de sistema externo",
                "description": "Endpoint para receber webhooks de sistemas externos com validação de segurança",
                "operationId": "receiveWebhook",
                "requestBody": {
                    "description": "Payload do webhook",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "data"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Tipo do evento",
                                        "type": "string",
                                        "example": "payment.completed"
                                    },
                                    "data": {
                                        "description": "Dados do evento",
                                        "type": "object"
                                    },
                                    "timestamp": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-11-27T10:00:00Z"
                                    },
                                    "signature": {
                                        "description": "Assinatura HMAC do webhook",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook received and queued for processing"
                                        },
                                        "webhook_id": {
                                            "type": "string",
                                            "example": "wh_1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid webhook signature"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Muitas requisições",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too many requests"
                                        },
                                        "retry_after": {
                                            "type": "integer",
                                            "example": 60
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/webhooks/send": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Enviar webhook para sistema externo",
                "description": "Envia webhook de forma assíncrona usando filas Redis. Suporta priorização e retry automático.",
                "operationId": "sendWebhook",
                "requestBody": {
                    "description": "Dados do webhook a ser enviado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url",
                                    "payload"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/webhook"
                                    },
                                    "payload": {
                                        "description": "Dados a serem enviados",
                                        "type": "object"
                                    },
                                    "priority": {
                                        "description": "Prioridade de processamento",
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "normal",
                                            "high",
                                            "critical",
                                            "webhook"
                                        ],
                                        "example": "normal"
                                    },
                                    "headers": {
                                        "description": "Headers customizados",
                                        "type": "object",
                                        "example": {
                                            "X-Custom-Header": "value"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Webhook enfileirado para envio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook queued for delivery"
                                        },
                                        "priority": {
                                            "type": "string",
                                            "example": "normal"
                                        },
                                        "estimated_delivery": {
                                            "type": "string",
                                            "example": "within 30 seconds"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/webhooks/send-sync": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Enviar webhook síncrono (bloqueante)",
                "description": "Envia webhook de forma síncrona. Use apenas para testes ou casos críticos onde resposta imediata é necessária.",
                "operationId": "sendWebhookSync",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url",
                                    "payload"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "payload": {
                                        "type": "object"
                                    },
                                    "timeout": {
                                        "description": "Timeout em segundos",
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "headers": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook enviado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "body": {
                                            "type": "object"
                                        },
                                        "duration_ms": {
                                            "type": "number",
                                            "example": 145.67
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao enviar webhook",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via Bearer Token (Sanctum)",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}